+++ Sources/SourceKitD/DynamicallyLoadedSourceKitD.swift
@@ -12,16 +12,10 @@
 
 import Foundation
 import LSPLogging
-import SKSupport
 import SwiftExtensions
 
 import struct TSCBasic.AbsolutePath
 
-#if compiler(<5.11)
-extension DLHandle: @unchecked Sendable {}
-#else
-extension DLHandle: @unchecked @retroactive Sendable {}
-#endif
 extension sourcekitd_api_keys: @unchecked Sendable {}
 extension sourcekitd_api_requests: @unchecked Sendable {}
 extension sourcekitd_api_values: @unchecked Sendable {}
