+++ Sources/SwiftExtensions/PipeAsStringHandler.swift
