+++ Sources/SwiftExtensions/ResultExtensions.swift
