+++ Tests/SourceKitLSPTests/BuildSystemTests.swift
@@ -15,6 +15,7 @@ import LSPTestSupport
 import LanguageServerProtocol
 @_spi(Testing) import SKCore
 import SKTestSupport
+@_spi(Testing) import SemanticIndex
 @_spi(Testing) import SourceKitLSP
 import TSCBasic
 import XCTest
